+++ packages/grafana-data/src/index.ts
@@ -435,7 +435,7 @@ export {
   isStandardFieldProp,
   type OptionDefaults,
 } from './panel/getPanelOptionsWithDefaults';
-export { getPanelDataSummary } from './panel/getPanelDataSummary';
+export { type PanelDataSummary, getPanelDataSummary } from './panel/getPanelDataSummary';
 export { createFieldConfigRegistry } from './panel/registryFactories';
 export { type QueryRunner, type QueryRunnerOptions } from './types/queryRunner';
 export { type GroupingToMatrixTransformerOptions } from './transformations/transformers/groupingToMatrix';
@@ -652,7 +652,6 @@ export {
   type AngularPanelMenuItem,
   type PanelPluginDataSupport,
   type VisualizationSuggestion,
-  type PanelDataSummary,
   type VisualizationSuggestionsSupplier,
   VizOrientation,
   VisualizationSuggestionScore,
